\name{fmri.gui}
\alias{fmri.gui}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Graphical user interface }
\description{ The function provides a graphical user interface that guides through
the analysis of single subject fmri analysis from assessing the image data to
visualization of results.
}
\usage{
fmri.gui()
}
%- maybe also 'usage' for other objects documented here.
%\details{
%}
\value{ Results of the analysis are stored in a file or saved in the global workspace.
}
\author{ Devy Hoffmann and Karsten Tabelow \email{tabelow@wias-berlin.de}  }
%\note{ }
\seealso{ \code{\link{read.AFNI}}, \code{\link{read.ANALYZE}}, \code{\link{read.DICOM}},  \code{\link{fmri.design}}, \code{\link{fmri.stimulus}},
 \code{\link{fmri.stimulus}}, \code{\link{fmri.lm}}, \code{\link{fmri.smooth}}, \code{\link{fmri.pvalue}}, \code{\link{plot.fmridata}},  \code{\link{print.fmridata}}, \code{\link{write.AFNI}},\code{\link{write.ANALYZE}},\code{\link{write.NIFTI}}
}
\examples{
\dontrun{fmri.gui()}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }
