% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fmridata.R
\name{gen_fmridata}
\alias{gen_fmridata}
\title{Generate fmridata example}
\usage{
gen_fmridata(signal = 1.5, noise = 20, arfactor = 0.3)
}
\arguments{
\item{signal}{Level of signal in the data}

\item{noise}{Level of noise in the data}

\item{arfactor}{Autoregressive factor}
}
\value{
Object of class \code{fmridata}
}
\description{
Generate fmridata example
}
\examples{
\donttest{
gen_fmridata()
}
}
