\name{hvred}
\alias{hvred}
\docType{data}
\title{
Translation between smoothness and bandwidth for Gaussian kernel
}
\description{
Translation table between smoothness and bandwidth for Gaussian kernel
}
\usage{data(hvred)}
\format{
  The format is:
 num [1:500, 1:2] 0.101 0.102 0.103 0.104 0.105 ...
}
\examples{
data(hvred)
## maybe str(hvred) ; plot(hvred) ...
}
\keyword{datasets}
